pyproject.toml
[tool.poetry]
name = "flask-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
flask = "^3.0.3"
python-dotenv = "^1.0.1"
click = "^8.1.7"
opencv-python = "^4.10.0.84"
pillow = "^10.4.0"
numpy = "^2.0.1"
mypy-extensions = "^1.0.0"
torch = "^2.4.0"
torchvision = "^0.19.0"
pytest = "^8.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

